Add new institutions package under tools - move OpsUser into its own file
