+++ api/build.gradle
@@ -825,7 +825,7 @@ task loadDataDictionary(type: JavaExec) {
 // See project.rb command: load-institutions
 task loadInstitutions(type: JavaExec) {
   classpath = sourceSets.__tools__.runtimeClasspath
-  main = "org.pmiops.workbench.tools.LoadInstitutions"
+  main = "org.pmiops.workbench.tools.institutions.LoadInstitutions"
   systemProperties = commandLineSpringProperties
   if (project.hasProperty("appArgs")) {
     args Eval.me(appArgs)
@@ -835,7 +835,7 @@ task loadInstitutions(type: JavaExec) {
 // See project.rb command: populate-ops-user-affiliations
 task populateOpsUserAffiliations(type: JavaExec) {
   classpath = sourceSets.__tools__.runtimeClasspath
-  main = "org.pmiops.workbench.tools.PopulateOpsUserAffiliations"
+  main = "org.pmiops.workbench.tools.institutions.PopulateOpsUserAffiliations"
   systemProperties = commandLineSpringProperties
   if (project.hasProperty("appArgs")) {
     args Eval.me(appArgs)
