+++ api/tools/src/main/java/org/pmiops/workbench/tools/institutions/LoadInstitutions.java
@@ -1,4 +1,4 @@
-package org.pmiops.workbench.tools;
+package org.pmiops.workbench.tools.institutions;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.BufferedReader;
@@ -15,6 +15,7 @@ import org.pmiops.workbench.institution.InstitutionServiceImpl;
 import org.pmiops.workbench.institution.InstitutionUserInstructionsMapperImpl;
 import org.pmiops.workbench.institution.PublicInstitutionDetailsMapperImpl;
 import org.pmiops.workbench.model.Institution;
+import org.pmiops.workbench.tools.CommandLineToolConfig;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
